fix: [PL-40064]: fix client call to delete membership (#51275)
* fix: [PL-40064]: fix client call to delete membership

+++ 400-rest/src/main/java/software/wings/service/impl/UserServiceImpl.java
@@ -3276,7 +3276,7 @@ public class UserServiceImpl implements UserService {
       List<Account> updatedActiveAccounts = userServiceHelper.updatedActiveAccounts(user, accountId);
       List<Account> updatedPendingAccounts = userServiceHelper.updatedPendingAccount(user, accountId);
 
-      if (isUserPartOfAccountInNG.get()) {
+      if (isUserPartOfAccountInNG.get() && removeUserFilter.equals(NGRemoveUserFilter.ACCOUNT_LAST_ADMIN_CHECK)) {
         userServiceHelper.deleteUserFromNG(userId, accountId, removeUserFilter);
       }
